# pkg/webui/handlers/docs/swagger.yaml
definitions:
  action_plan.BidPlan:
    properties:
      bid_end_time:
        type: integer
      bid_increase:
        description: gwei
        type: integer
      bid_interval:
        description: ms, >= 0, 0 = single bid
        type: integer
      bid_min_amount:
        description: gwei
        type: integer
      bid_start_time:
        type: integer
      bid_subsidy:
        description: gwei
        type: integer
      bid_value_gwei:
        description: |-
          BidValueGwei is an absolute bid base value replacing
          max(blockValue, min) + subsidy; BidIncrease still applies per re-bid.
          Allows underbidding the block value for testing.
        type: integer
      ignore_missing_prefs:
        description: |-
          IgnoreMissingPrefs bids with the payload's fee recipient when no gossip
          proposer preferences arrived for the slot, bypassing the skip gate.
        type: boolean
      mode:
        $ref: '#/definitions/action_plan.Mode'
    type: object
  action_plan.BuilderAPIPlan:
    properties:
      mode:
        $ref: '#/definitions/action_plan.Mode'
      response_delay_ms:
        description: |-
          ResponseDelayMs delays the bid response by this many milliseconds
          (context-cancellable, capped at one slot).
        type: integer
      total_value_override_gwei:
        description: |-
          TotalValueOverrideGwei is the absolute total proposer-visible bid value
          (before the Gloas execution-payment split), replacing block value +
          subsidy. May exceed the block value to test payment edge cases.
        type: integer
      value_subsidy_gwei:
        description: ValueSubsidyGwei replaces the global BlockValueSubsidyGwei for
          this slot.
        type: integer
    type: object
  action_plan.FrozenPlan:
    properties:
      bid:
        allOf:
        - $ref: '#/definitions/action_plan.ResolvedBidSettings'
        description: |-
          Resolved effective settings; a nil category = suppressed for this slot
          (by plan or by the global enable flags / protocol applicability).
      build:
        allOf:
        - $ref: '#/definitions/action_plan.ResolvedBuildSettings'
        description: |-
          Build is the complete resolved build decision for the slot (schedule +
          plan force/suppress). Always non-nil.
      builder_api:
        $ref: '#/definitions/action_plan.ResolvedBuilderAPISettings'
      fork:
        description: fork name at the target slot
        type: string
      frozen_at:
        type: string
      plan:
        allOf:
        - $ref: '#/definitions/action_plan.SlotPlan'
        description: raw sparse plan; nil = none existed
      reveal:
        $ref: '#/definitions/action_plan.ResolvedRevealSettings'
      slot:
        type: integer
    type: object
  action_plan.Mode:
    enum:
    - disabled
    - custom
    type: string
    x-enum-varnames:
    - ModeDisabled
    - ModeCustom
  action_plan.PlanUpdate:
    properties:
      bid:
        items:
          type: integer
        type: array
      builder_api:
        items:
          type: integer
        type: array
      delete:
        type: boolean
      from_slot:
        type: integer
      reveal:
        items:
          type: integer
        type: array
      set:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
      slots:
        items:
          type: integer
        type: array
      to_slot:
        type: integer
    type: object
  action_plan.ResolvedBidSettings:
    properties:
      end_ms:
        type: integer
      forced:
        description: |-
          Forced marks that the plan activated bidding although the module is
          globally disabled.
        type: boolean
      ignore_missing_prefs:
        description: IgnoreMissingPrefs bids without gossip proposer preferences.
        type: boolean
      increase_gwei:
        type: integer
      interval_ms:
        type: integer
      min_gwei:
        type: integer
      start_ms:
        type: integer
      subsidy_gwei:
        type: integer
      value_gwei:
        description: |-
          ValueGwei, when set, is the absolute bid base value replacing the
          max(blockValue, min) + subsidy formula (IncreaseGwei still applies).
        type: integer
    type: object
  action_plan.ResolvedBuildSettings:
    properties:
      build:
        description: 'Build is the final decision: build a payload for this slot or
          not.'
        type: boolean
      build_start_time_ms:
        description: |-
          BuildStartTimeMs is the effective build start time, milliseconds
          relative to slot start (signed).
        type: integer
      forced:
        description: |-
          Forced marks builds the plan pushed past the schedule (they never
          consume the next_n budget).
        type: boolean
      plan_involved:
        description: |-
          PlanInvolved marks decisions where a per-slot plan existed or any
          consumer was effectively active — i.e. skips worth surfacing.
        type: boolean
      skip_reason:
        description: |-
          SkipReason is one of the BuildSkipReason* constants when Build is
          false, empty otherwise.
        type: string
    type: object
  action_plan.ResolvedBuilderAPISettings:
    properties:
      delay_ms:
        type: integer
      forced:
        description: |-
          Forced marks that the plan activated serving although the module is
          globally disabled.
        type: boolean
      subsidy_gwei:
        type: integer
      total_value_gwei:
        description: |-
          TotalValueGwei, when set, is the absolute total proposer-visible bid
          value (before the Gloas execution-payment split).
        type: integer
    type: object
  action_plan.ResolvedRevealSettings:
    properties:
      bypass_deadline:
        description: |-
          BypassDeadline disables the "past the in-slot deadline → skip" check so
          deliberately late reveals are attempted.
        type: boolean
      reveal_time_ms:
        type: integer
      suppressed:
        type: boolean
    type: object
  action_plan.RevealPlan:
    properties:
      mode:
        $ref: '#/definitions/action_plan.Mode'
      reveal_time_ms:
        description: RevealTimeMs is signed milliseconds relative to slot start.
        type: integer
    type: object
  action_plan.SlotPlan:
    properties:
      bid:
        $ref: '#/definitions/action_plan.BidPlan'
      builder_api:
        $ref: '#/definitions/action_plan.BuilderAPIPlan'
      reveal:
        $ref: '#/definitions/action_plan.RevealPlan'
      slot:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  api.ActionPlanResponse:
    properties:
      max_slot:
        type: integer
      min_slot:
        type: integer
      plans:
        items:
          $ref: '#/definitions/action_plan.SlotPlan'
        type: array
    type: object
...
  api.SlotResultsResponse:
    properties:
      max_slot:
        type: integer
      min_slot:
        type: integer
      results:
        items:
          $ref: '#/definitions/slot_results.SlotResult'
        type: array
    type: object
...
  api.UpdateActionPlanRequest:
    properties:
      updates:
        items:
          $ref: '#/definitions/action_plan.PlanUpdate'
        type: array
    type: object
  api.UpdateActionPlanResponse:
    properties:
      plans:
        items:
          $ref: '#/definitions/action_plan.SlotPlan'
        type: array
      slots:
        items:
          type: integer
        type: array
      status:
        type: string
    type: object
...
  slot_results.BidAttempt:
    properties:
      artifact_index:
        description: |-
          ArtifactIndex references the slot's stored 'bid' SSZ artifact; nil when
          no artifact exists (suppressed, pre-construction failure, or capture
          disabled/failed).
        type: integer
      at:
        type: string
      competitor_high_gwei:
        description: |-
          CompetitorHighGwei is the highest known competitor bid at submission
          time (p2p only, our own builder index excluded).
        type: integer
      error:
        type: string
      execution_payment_gwei:
        type: integer
      status:
        $ref: '#/definitions/slot_results.BidStatus'
      total_value_gwei:
        type: integer
      transport:
        description: payload_builder.BidTransport values
        type: string
    type: object
  slot_results.BidStatus:
    enum:
    - suppressed
    - constructed
    - submitted
    - served
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - BidStatusSuppressed
    - BidStatusConstructed
    - BidStatusSubmitted
    - BidStatusServed
    - BidStatusFailed
    - BidStatusCancelled
  slot_results.BlockSubmission:
    properties:
      at:
        type: string
      dialect:
        description: '"legacy" | "epbs"'
        type: string
      error:
        type: string
      status:
        $ref: '#/definitions/slot_results.SubmissionStatus'
    type: object
  slot_results.BuildOutcome:
    properties:
      at:
        type: string
      block_hash:
        type: string
      block_value_wei:
        type: string
      error:
        type: string
      fee_recipient:
        type: string
      num_blobs:
        type: integer
      num_transactions:
        type: integer
      skip_reason:
        description: action_plan.BuildSkipReason* when skipped
        type: string
      status:
        $ref: '#/definitions/slot_results.BuildStatus'
    type: object
  slot_results.BuildStatus:
    enum:
    - waiting_attributes
    - no_attributes
    - started
    - ready
    - failed
    - skipped
    type: string
    x-enum-varnames:
    - BuildStatusWaitingAttributes
    - BuildStatusNoAttributes
    - BuildStatusStarted
    - BuildStatusReady
    - BuildStatusFailed
    - BuildStatusSkipped
  slot_results.InclusionResult:
    properties:
      block_hash:
        type: string
      num_blobs:
        type: integer
      num_transactions:
        type: integer
      source:
        description: '"epbs" | "builder_api"'
        type: string
      timestamp:
        type: string
      value_eth:
        type: string
      value_wei:
        type: string
    type: object
  slot_results.RevealAttempt:
    properties:
      at:
        type: string
      attempt:
        type: integer
      error:
        type: string
      skip_reason:
        description: plan_disabled | late
        type: string
      status:
        $ref: '#/definitions/slot_results.RevealStatus'
      transport:
        type: string
    type: object
  slot_results.RevealStatus:
    enum:
    - suppressed
    - published
    - failed
    - skipped
    type: string
    x-enum-varnames:
    - RevealStatusSuppressed
    - RevealStatusPublished
    - RevealStatusFailed
    - RevealStatusSkipped
  slot_results.SlotResult:
    properties:
      applied_plan:
        allOf:
        - $ref: '#/definitions/action_plan.FrozenPlan'
        description: AppliedPlan is the frozen plan snapshot the slot executed under.
      bids:
        items:
          $ref: '#/definitions/slot_results.BidAttempt'
        type: array
      block_submissions:
        items:
          $ref: '#/definitions/slot_results.BlockSubmission'
        type: array
      build:
        $ref: '#/definitions/slot_results.BuildOutcome'
      dropped_attempts:
        additionalProperties:
          type: integer
        description: |-
          DroppedAttempts counts attempts beyond the per-kind retention cap,
          keyed by kind ("bids", "block_submissions", "reveal_attempts").
        type: object
      epoch:
        type: integer
      fork:
        type: string
      inclusion:
        $ref: '#/definitions/slot_results.InclusionResult'
      reveal_attempts:
        items:
          $ref: '#/definitions/slot_results.RevealAttempt'
        type: array
      slot:
        type: integer
      updated_at:
        type: string
    type: object
  slot_results.SubmissionStatus:
    enum:
    - received
    - accepted
    - failed
    type: string
    x-enum-varnames:
    - SubmissionStatusReceived
    - SubmissionStatusAccepted
    - SubmissionStatusFailed
info:
  contact: {}
paths:
  /api/buildoor/action-plan:
    get:
      description: Returns all per-slot action plans within the inclusive slot range.
      operationId: getActionPlan
      parameters:
      - description: Range start slot (inclusive)
        in: query
        name: min_slot
        required: true
        type: integer
      - description: Range end slot (inclusive)
        in: query
        name: max_slot
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ActionPlanResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Plan service unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get per-slot action plans
      tags:
      - ActionPlan
    post:
      consumes:
      - application/json
      description: |-
        Applies a bulk per-slot action plan mutation atomically (all
        targeted slots update or none). Category members are
        three-state (absent = unchanged, null = clear, object =
        replace); the "set" member supports fine-grained path updates
        like {"bid.bid_min_amount": 5000} for partial edits. Updates
        targeting past or already-frozen slots fail with 409.
        Requires authentication.
      operationId: updateActionPlan
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Plan updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateActionPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authoritative normalized result
          schema:
            $ref: '#/definitions/api.UpdateActionPlanResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Slot in the past or frozen
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Plan service unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update per-slot action plans
      tags:
      - ActionPlan
...
  /api/buildoor/slot-results:
    get:
      description: |-
        Returns the recorded outcome history (build, bids, block
        submissions, reveals, inclusion, applied plan) for every
        active slot within the inclusive slot range.
      operationId: getSlotResults
      parameters:
      - description: Range start slot (inclusive)
        in: query
        name: min_slot
        required: true
        type: integer
      - description: Range end slot (inclusive)
        in: query
        name: max_slot
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SlotResultsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Results tracker unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get per-slot results
      tags:
      - ActionPlan
...
  /api/config/settings:
    post:
      consumes:
      - application/json
      description: |-
        Applies partial global settings updates keyed by canonical
        settings paths (e.g. {"epbs.bid_subsidy": 1000,
        "schedule.mode": "all"}) without requiring full config
        objects. Unknown keys are rejected. Requires authentication.
      operationId: updateSettings
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map of settings key paths to values
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Unknown key or invalid value
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update global settings by key path
      tags:
      - Config
...
swagger: "2.0"